Adds user registration screens and forms
Introduces separate registration screens and forms for regular users and auctioneers.

Implements form validation using React Hook Form and Zod.

Integrates with the `useSignup` hook to handle user registration and account creation.

Adds navigation to the login screen after successful registration.

Removes Facebook auth provider.
