+++ src/hooks/auth/useSignup.ts
@@ -6,55 +6,15 @@
 import { useState, useCallback } from 'react';
 import { useSecureApi } from '../api/useSecureApi';
 import { PROTECTED_ENDPOINTS } from '@/config/api-config';
-import type { Lang, LangMap, UserRoles } from '@/types/types';
+import type {
+  Lang,
+  LangMap,
+  UserRoles,
+  SignupData,
+  SignupResponse,
+  UseSignupReturn,
+} from '@/types/types';
 import * as Sentry from '@sentry/react-native';
-
-export interface SignupData {
-  // Campos comunes para todos los roles
-  name: string;
-  lastName: string;
-  username: string;
-  email: string;
-  password: string;
-  confirmPassword: string;
-
-  // Campos opcionales según el rol
-  dni?: string;
-  phoneNumber?: string;
-  profilePicture?: string;
-
-  // Campos adicionales para AUCTIONEER
-  storeName?: string;
-  webPage?: string;
-  socialMedia?: string;
-  address?: string;
-  town?: string;
-  province?: string;
-  country?: string;
-  postalCode?: string;
-}
-
-interface SignupResponse {
-  error: LangMap | null;
-  success: LangMap | null;
-  data: {
-    email: string;
-  } | null;
-}
-
-interface UseSignupReturn {
-  signup: (
-    data: SignupData,
-    role: UserRoles,
-    lang: Lang
-  ) => Promise<{
-    success: boolean;
-    email?: string;
-    error?: LangMap;
-  }>;
-  isLoading: boolean;
-  errorMessage: LangMap | null;
-}
 
 export const useSignup = (): UseSignupReturn => {
   const [isLoading, setIsLoading] = useState(false);
