+++ src/types/types.ts
@@ -732,3 +732,51 @@ export type SubscribeStatus =
   | 'TIMED_OUT'
   | 'CLOSED'
   | 'CHANNEL_ERROR';
+
+// Auth - Signup Types
+export interface SignupData {
+  // Campos comunes para todos los roles
+  name: string;
+  lastName: string;
+  username: string;
+  email: string;
+  password: string;
+  confirmPassword: string;
+
+  // Campos opcionales según el rol
+  dni?: string;
+  phoneNumber?: string;
+  profilePicture?: string;
+
+  // Campos adicionales para AUCTIONEER
+  storeName?: string;
+  webPage?: string;
+  socialMedia?: string;
+  address?: string;
+  town?: string;
+  province?: string;
+  country?: string;
+  postalCode?: string;
+}
+
+export interface SignupResponse {
+  error: LangMap | null;
+  success: LangMap | null;
+  data: {
+    email: string;
+  } | null;
+}
+
+export interface UseSignupReturn {
+  signup: (
+    data: SignupData,
+    role: UserRoles,
+    lang: Lang
+  ) => Promise<{
+    success: boolean;
+    email?: string;
+    error?: LangMap;
+  }>;
+  isLoading: boolean;
+  errorMessage: LangMap | null;
+}
